requirements.txt
absl-py==1.0.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astor==0.8.1
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==5.0.1
blessings==1.7
bs4==0.0.1
cachetools==4.2.4
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==2.0.9
click==8.0.4
cloudpickle==3.0.0
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.0.7
crypto==1.4.1
cvxopt==1.3.0
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dgl==1.1.3+cu116
distlib==0.3.8
Django==4.1.3
django-cors-headers==3.13.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.13.1
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.3.3
geatpy==2.7.0
gensim==4.2.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==0.6.0
GraKeL==0.1.9
greenlet==3.0.3
grpcio==1.62.1
h5py==2.10.0
huggingface-hub==0.10.1
humanfriendly==10.0
idna==3.7
importlib-metadata==4.8.2
importlib-resources==5.10.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kaleido==0.2.1
keras==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lxml==4.7.1
Mako==1.3.5
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.5
Naked==0.1.32
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.0
numpy==1.22.0
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.103
nvidia-cuda-nvrtc-cu12==12.3.103
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.6.50
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.101
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-dali-cuda110==1.23.0
nvidia-dali-nvtf-plugin==1.23.0+nv23.3
nvidia-horovod==0.27.0+nv23.3
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-pyindex==1.0.9
nvidia-tensorflow==1.15.5+nv23.3
oauthlib==3.1.1
opt-einsum==3.3.0
optuna==3.6.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.20.0
pyg-lib @ file:///home/inspur/wzh/flrec/env/pyg_lib-0.2.0%2Bpt113cu116-cp38-cp38-linux_x86_64.whl#sha256=f77b357bcf3a89c62b0ed619e433934d333b471052b458516df80759a29f595e
Pygments==2.13.0
pyparsing==3.0.7
pyrsistent==0.19.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.3.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.1.3
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.96
setproctitle==1.3.3
shellescape==3.8.1
six==1.16.0
sklearn==0.0
smart-open==6.2.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.31
sqlparse==0.4.3
stack-data==0.6.0
style==1.1.0
tensorboard==1.15.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow-estimator==1.15.1
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
terminado==0.17.0
threadpoolctl==3.0.0
tinycss2==1.2.1
tokenizers==0.11.6
torch==1.12.1+cu116
torch-cluster @ file:///home/inspur/wzh/flrec/env/torch_cluster-1.6.0%2Bpt113cu116-cp38-cp38-linux_x86_64.whl#sha256=1523d1f23ef78cfca6af00dd36d2c5fad9a47b434e0937545e1aeb13f33bb57e
torch-scatter @ file:///home/inspur/wzh/flrec/env/torch_scatter-2.0.9-cp38-cp38-linux_x86_64.whl#sha256=ae2b2e666c8fa8383d639694463fff53f1e06ba66bdaa6b6cdd8a4aebe3f025a
torch-sparse @ file:///home/inspur/wzh/flrec/env/torch_sparse-0.6.15%2Bpt113cu116-cp38-cp38-linux_x86_64.whl#sha256=79b44e7a0d75967b41982822ffa76c8deb15269ea7b2697848d2cce96d289e61
torch-spline-conv @ file:///home/inspur/wzh/flrec/env/torch_spline_conv-1.2.1%2Bpt113cu116-cp38-cp38-linux_x86_64.whl#sha256=342a762805a68e022f70faaed576b47350dbd7c2c05715ac3926de59b8e00b44
torch_geometric==2.5.0
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.62.3
traitlets==5.5.0
transformers==4.24.0
typing_extensions==4.12.2
update==0.0.1
urllib3==1.26.19
virtualenv==20.25.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.0.2
widgetsnbextension==4.0.3
wrapt==1.16.0
wsdream==1.0
XlsxWriter==3.0.3
yarl==1.9.4
zipp==3.6.0
